+++ palettes/gaming/superMarioBros/src/index.ts
@@ -1,0 +1,14 @@
+import { type Engine } from "@tsparticles/engine";
+import { options } from "./options.js";
+
+const paletteName = "super-mario-bros";
+
+/**
+ *
+ * @param engine
+ */
+export async function loadSuperMarioBrosPalette(engine: Engine): Promise<void> {
+  await engine.pluginManager.register(e => {
+    e.pluginManager.addPalette(paletteName, options);
+  });
+}